1. build-openwrt.sh: 'make V=99' for each time; 2. tw150v1-buildfw.sh: added check to the prefix image.

+++ trunk/firmware-builder/tw150v1-buildfw.sh
@@ -12,23 +12,37 @@ if [ ! -f prefix-1-2.bin ]; then
 	( set -x; head "$official_fw" -c131072 > prefix-1-2.bin )
 fi
 
-set -x
+# Check if it is a valid prefix image or truncated
+prefix_size=`stat --format=%s prefix-1-2.bin`
+if [ $prefix_size -ne 131072 ]; then
+	echo "*** prefix-1-2.bin has invalid size. Please remove it and 'svn up'."
+	exit 1
+fi
 
-cp -vf $openwrt_root/build_dir/target-mips*/linux-ar71xx_generic/vmlinux   vmlinux-hiwifi-tw150v1
-$openwrt_root/staging_dir/host/bin/patch-cmdline vmlinux-hiwifi-tw150v1   "board=tw150v1 console=ttyATH0,115200"
-$openwrt_root/staging_dir/host/bin/lzma   e vmlinux-hiwifi-tw150v1 -lc1 -lp2 -pb2 vmlinux-hiwifi-tw150v1.bin.lzma
+(
+	set -x
 
-cp $openwrt_root/bin/ar71xx/$root_squashfs ./
+	cp -vf $openwrt_root/build_dir/target-mips*/linux-ar71xx_generic/vmlinux   vmlinux-hiwifi-tw150v1
+	$openwrt_root/staging_dir/host/bin/patch-cmdline vmlinux-hiwifi-tw150v1   "board=tw150v1 console=ttyATH0,115200"
+	$openwrt_root/staging_dir/host/bin/lzma   e vmlinux-hiwifi-tw150v1 -lc1 -lp2 -pb2 vmlinux-hiwifi-tw150v1.bin.lzma
 
-mkimage -A mips -O linux -T kernel -a 0x80060000 -C lzma  -e 0x80060000 -n 'tw150v1-Linux-3.8.3' -d vmlinux-hiwifi-tw150v1.bin.lzma uImage
+	cp $openwrt_root/bin/ar71xx/$root_squashfs ./
 
-raw_size=`stat --format="%s" uImage`
-pad_size=`expr 1310720 - $raw_size`
-dd if=/dev/zero of=__pad.tmp bs=$pad_size count=1
-cat uImage __pad.tmp > uImage.1280k
-rm -f __pad.tmp
+	mkimage -A mips -O linux -T kernel -a 0x80060000 -C lzma  -e 0x80060000 -n 'tw150v1-Linux-3.8.3' -d vmlinux-hiwifi-tw150v1.bin.lzma uImage
 
-cat prefix-1-2.bin uImage.1280k $root_squashfs > $target_fw
+	raw_size=`stat --format=%s uImage`
+	pad_size=`expr 1310720 - $raw_size`
+	dd if=/dev/zero of=__pad.tmp bs=$pad_size count=1
+	cat uImage __pad.tmp > uImage.1280k
+	rm -f __pad.tmp
 
-cp -vf $target_fw /tftpboot/recovery.bin
+	cat prefix-1-2.bin uImage.1280k $root_squashfs > $target_fw
+) || exit $?
+
+[ -d tftpboot ] && cp -vf $target_fw /tftpboot/recovery.bin || :
+
+echo ""
+echo "###############################################"
+echo "Target firmware: $target_fw"
+echo "###############################################"
